Move ServerNetworking.{h,cpp} into server directory
