+++ msvc2017/FreeOrionD/FreeOrionD.vcxproj.filters
@@ -83,9 +83,6 @@
     <ClInclude Include="..\..\network\Networking.h">
       <Filter>Header Files\network</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\network\ServerNetworking.h">
-      <Filter>Header Files\network</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\universe\Building.h">
       <Filter>Header Files\universe</Filter>
     </ClInclude>
@@ -156,6 +153,9 @@
       <Filter>Header Files\server</Filter>
     </ClInclude>
     <ClInclude Include="..\..\server\ServerFSM.h">
+      <Filter>Header Files\server</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\server\ServerNetworking.h">
       <Filter>Header Files\server</Filter>
     </ClInclude>
     <ClInclude Include="..\..\server\ServerWrapper.h">
@@ -236,9 +236,6 @@
     <ClInclude Include="StdAfx.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\network\ServerNetworking.cpp">
-      <Filter>Source Files\network</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\server\ServerFSM.cpp">
       <Filter>Source Files\server</Filter>
     </ClCompile>
@@ -249,6 +246,9 @@
       <Filter>Source Files\server</Filter>
     </ClCompile>
     <ClCompile Include="..\..\server\ServerFramework.cpp">
+      <Filter>Source Files\server</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\server\ServerNetworking.cpp">
       <Filter>Source Files\server</Filter>
     </ClCompile>
     <ClCompile Include="..\..\server\ServerWrapper.cpp">
