+++ network/CMakeLists.txt
@@ -8,10 +8,3 @@ target_sources(freeorioncommon
         ${CMAKE_CURRENT_LIST_DIR}/MessageQueue.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Networking.cpp
 )
-
-target_sources(freeoriond
-    PUBLIC
-        ${CMAKE_CURRENT_LIST_DIR}/ServerNetworking.h
-    PRIVATE
-        ${CMAKE_CURRENT_LIST_DIR}/ServerNetworking.cpp
-)
